Move sub components into a separate file. Updated tests to reflect this change
