+++ packages/components/src/Modal/index.ts
@@ -1,1 +1,2 @@
+export { useModalContext } from "./ModalContext";
 export { Modal } from "./Modal";
